PID Gains Optimization using PSO
This is a part of in-class project submission for IC Engine Idle Speed PID Gains Optimization for minimal cost. Consider this project as an engineering application to the optimization algorithm. Initially the Simulink is used to optimize manually the tuner, then FMINCON is used to further optimize the results. 
Then I used PSO and Pattern Search Optimization (intuitions provided in the report) for global optimization and further identified the best parameters.